GH-203: [refactor] Forgot to add to last commit, files with changes that followed the directory changes, minor changes
