+++ ClientApp/app/auth/registerAccount.tsx
@@ -6,8 +6,8 @@ import { useRouter } from "expo-router";
 import { IconPlacement } from '@/utils/constants/enums';
 import { hs, mhs, mvs, vs } from "@/utils/helpers/uiScaler";
 import { loginUser, registerUser } from "@/services/authService";
-import ConfirmButton from "@/components/ConfirmButton";
-import AuthenticationDivider from "@/components/AuthenticationDivider";
+import ConfirmButton from "@/components/Helper Components/ConfirmButton";
+import AuthenticationDivider from "@/components/Helper Components/AuthenticationDivider";
 import Checkbox from 'expo-checkbox';
 import themeColors from "@/utils/constants/colors";
 import { useUpdateUserToStore } from "@/state/user/actions";
